added Roland´s suggestions
